feat: Add weapon pick up system.
Added a system that player can pick up the weapon if the player does not have a type of the weapon.
This process is synchronized, so other players can see it.

TODO: When a new player connects after equip the weapon, the new player's screen does not properly represent the flip of the weapon sprite in the direction of the mouse. need to fix it.