fix: Fix weapon double attack issue
Fixed an issue where the first shot of a weapon fired two bullets at a time.

The firing mechanism of the gun has been optimized.

I'm working on an issue where different players appear to have acquired a weapon when multiple players rub against each other on a dropped weapon.

Resolves: #a84e8b4 - third issue.