feat: Add player dead animation.
I'm working on player death processing.
This will include the act of throwing away weapons, the addition of viewing modes, and the act of investigating dead players.

After a dead player leaves the game, when a new player comes in, the dead player's body is invisible (because the RPC of the player who left is not called) and the location of the abandoned weapon and body are not properly synchronized.
I need to fix them.