fix: Fix the issue of body location synchronization
Fixed an issue where the position of the dead player's corpse would not sync properly when a new player entered the game after the previous player had died.

When a new player joins, the dead player's death animation is playing at the same time. However, I decided that this bug does not belong to an issue that is fatal to gameplay, so I plan to focus more on implementing existing bugs and features.

Resolves: #1485f54