fix: Fix game state synchronization when new player join
Fixed an issue where the state of the game was not properly synchronized when a new player joined after the master client left.
To resolve this issue, game states are now synchronized through custom properties rather than RPCs.

I'm currently continuing to work on the overall rules of the game and the game processing of the player's death.