feat: Add Player info UI
Added UI to display name and HP above the player. The color of the player name looks different depending on the role. Instead, the player who plays the role of the mafia appears to be a civil on the screen of civil, including detectives.

TODO: If a spectator comes in during the game, the color of the name of the existing player will not be displayed properly.
Also need to apply the UI of the Game Panel in the same way as the color of the player name (because of optimization).
Finally, the player UI should only display the player's information when a local player place the mouse on top of the player.