fix: fix player nick name color issue
Fixed an issue where the player's name color was not displayed properly when the spectator entered the room during the gameplay.
In addition, the mechanism of the Game Panel was adjusted for optimization.

TODO: Living players should not be allowed to see ghosts of dead players.
Need to implement a feature that allows the Player UI information to be displayed only when the mouse is placed on the player.

Resolves: #fbaedb2