feat: Add Mouse Enter Event
Mouse Enter Event has been added to display player information when the mouse is brought to the player.
In addition, the collision judgment of objects such as bullets was optimized by changing it to a layer check method.

TODO:  Need to fix the issue that the sprite of a previously dead player looks like a live player, not a ghost when a player becomes a ghost.