fix: Fix previously dead player sprite issue
Fixed an issue where the sprite of a previously dead player looked like a living player, not a ghost.

TODO: Need to fix an issue where the position of the weapon is not properly synchronized when a new player comes in after the player drops his weapon and leaves the game.

Resolves: #6cbf29f