feat: Synchronize ScoreBoard UI
I have completed synchronization so that will be applied to the scoreboard immediately after each change in the player's data. This applies well even if the player leaves the game or a new player joins.

TODO: I still need to do some extra work to show the player's ping, prevent the Civil player from seeing the Missing player, and hide the categories that have non player.