feat: Implement menu button
Added a menu to leave a room by pressing ESC during the game.

Currently, when the player exits the game, they reconnect to the Lobby Scene through the LogIn Scene, but need to improve this so that they can skip this process and go directly to the Lobby Scene in the future.

Also fixed some issues that the Room List not updating properly in the Lobby and the text outline overflowing in-game.

TODO:
- implement Player's sight and shadow
- implement room option for master
- implement room passwrod