feat: Implement room option for master client
Added an option in the menu for the master client to set the role.
Also added a UI to set the role when creating a new room in the lobby.
After a master client leaves the game, the new master client that takes over can also set up the game.

TODO:
- implement Player's sight and shadow
- implement room passwrod