feat: Implement body search on mobile
Added a button for searching bodies on mobile devices.

Modified body search mechanics for improved mobile functionality.

Instead of relying on OnMouse and OnTrigger for the active body search window in response to keyboard input at each body, the logic has been updated to determine the nearest body to the player and the body currently under the mouse cursor within the Map Manager.

This functionality was added to the Map Manager since it is responsible for spawning bodies. Additional managers may be required in the future to handle this functionality.

On mobile, the search targets the body closest to the player instead of relying on mouse input.